Move process check to its own package
