+++ pkg/compliance/resources/process/process_check_test.go
@@ -2,7 +2,8 @@
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
-package checks
+
+package process
 
 import (
 	"testing"
